Week6/requirements.txt
diffusers==0.27.2
facenet_pytorch==2.5.3
librosa==0.10.1
matplotlib==3.8.4
numpy==1.23.4
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.4.2
seaborn==0.13.2
torch==2.2.2
torchinfo==1.8.0
torchview==0.2.6
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
